# src/main/java/com/invillia/acme/dto/common/ErrorDTO.java
package com.invillia.acme.dto.common;

public class ErrorDTO {

    public String error;

    public String getError() {
        return error;
    }

    public void setError(String error) {
        this.error = error;
    }
    
}