# swagger-docker/dist/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "ACME project is a MVP project",
    "version" : "v1",
    "title" : "ACME Rest Documentation",
    "termsOfService" : "http://invillia.com/terms",
    "contact" : {
      "name" : "Support",
      "url" : "http://invillia.com.br",
      "email" : "support@invillia.com"
    },
    "license" : {
      "name" : "Invillia License 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
    "/api/order" : {
      "post" : {
        "tags" : [ "Order Services" ],
        "summary" : "",
        "description" : "Save order",
        "operationId" : "saveOrder",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/OrderInputDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/OrderOutputDTO"
            }
          },
          "201" : {
            "description" : "Order successfully saved",
            "schema" : {
              "$ref" : "#/definitions/OrderOutputDTO"
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          },
          "404" : {
            "description" : "Store not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/order/getById/{order-id}" : {
      "get" : {
        "tags" : [ "Order Services" ],
        "summary" : "",
        "description" : "Get order by id",
        "operationId" : "getOrderById",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "order-id",
          "in" : "path",
          "description" : "Order Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns the order",
            "schema" : {
              "$ref" : "#/definitions/OrderOutputDTO"
            }
          },
          "404" : {
            "description" : "The order id was not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/order/refund/{order-id}" : {
      "post" : {
        "tags" : [ "Order Services" ],
        "summary" : "",
        "description" : "Refund order",
        "operationId" : "refundOrder",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "order-id",
          "in" : "path",
          "description" : "Order Id",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Order successfully refunded",
            "schema" : {
              "$ref" : "#/definitions/MessageDTO"
            }
          },
          "403" : {
            "description" : "Order not allowed to be refunded",
            "schema" : {
              "$ref" : "#/definitions/MessageDTO"
            }
          },
          "404" : {
            "description" : "The order id was not found",
            "schema" : {
              "$ref" : "#/definitions/ErrorDTO"
            }
          }
        }
      }
    },
    "/api/payment" : {
      "post" : {
        "tags" : [ "Payment Services" ],
        "summary" : "",
        "description" : "Save payment",
        "operationId" : "savePayment",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/PaymentInputDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PaymentOutputDTO"
            }
          },
          "201" : {
            "description" : "Payment successfully saved",
            "schema" : {
              "$ref" : "#/definitions/PaymentOutputDTO"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "ErrorDTO" : {
      "type" : "object",
      "properties" : {
        "error" : {
          "type" : "string"
        }
      }
    },
    "MessageDTO" : {
      "type" : "object",
      "properties" : {
        "message" : {
          "type" : "string"
        }
      }
    },
    "OrderInputDTO" : {
      "type" : "object",
      "properties" : {
        "storeId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "address" : {
          "type" : "string"
        },
        "confirmationDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "PENDING", "AWAITING_PAYMENT", "AWAITING_FULFILLMENT", "AWAITING_SHIPMENT", "AWAITING_PICKUP", "PARTIALLY_SHIPPED", "COMPLETED", "SHIPPED", "CANCELLED", "DECLINED", "REFUNDED", "DISPUTED", "MANUAL_VERIFICACTION_REQUIRED", "PARTIALLY_REFUNDED" ]
        },
        "orderId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OrderItemInputDTO"
          }
        }
      }
    },
    "OrderItemInputDTO" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "unitPrice" : {
          "type" : "number",
          "format" : "float"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "OrderItemOutputDTO" : {
      "type" : "object",
      "properties" : {
        "description" : {
          "type" : "string"
        },
        "unitPrice" : {
          "type" : "number",
          "format" : "float"
        },
        "quantity" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "OrderOutputDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "storeId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "address" : {
          "type" : "string"
        },
        "confirmationDate" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "items" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/OrderItemOutputDTO"
          }
        }
      }
    },
    "PaymentInputDTO" : {
      "type" : "object",
      "properties" : {
        "orderId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "CANCELED_REVERSAL", "COMPLETED", "CREATED", "DENIED", "EXPIRED", "FAILED", "PENDING", "REFUNDED", "REVERSED", "PROCESSED", "VOIDED" ]
        },
        "creditCardNumber" : {
          "type" : "integer",
          "format" : "int64"
        },
        "paymentDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
    "PaymentOutputDTO" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int64"
        },
        "orderId" : {
          "type" : "integer",
          "format" : "int64"
        },
        "status" : {
          "type" : "string",
          "enum" : [ "CANCELED_REVERSAL", "COMPLETED", "CREATED", "DENIED", "EXPIRED", "FAILED", "PENDING", "REFUNDED", "REVERSED", "PROCESSED", "VOIDED" ]
        },
        "creditCardNumber" : {
          "type" : "integer",
          "format" : "int64"
        },
        "paymentDate" : {
          "type" : "string",
          "format" : "date-time"
        }
      }
    },
...
  }
}